started MovieDetail hooks
