+++ src/App.js
@@ -3,7 +3,6 @@ import { BrowserRouter as Router, Route, Switch, Link } from 'react-router-dom';
 import styled from 'styled-components';
 import MoviesList from './components/MoviesList';
 import MovieDetail from './components/MovieDetail';
-// import PracticeHook from './components/PracticeHook';
 
 const App = () => (
 	<Router>
