+++ src/components/MovieDetail.js
@@ -1,5 +1,6 @@
-import React, { PureComponent } from 'react';
+import React, { useState, useEffect } from 'react';
 import styled from 'styled-components';
+import axios from 'axios';
 import { Link } from 'react-router-dom';
 import { Poster } from './Movie';
 import Overdrive from 'react-overdrive';
@@ -7,47 +8,92 @@ import Overdrive from 'react-overdrive';
 const POSTER_PATH = 'http://image.tmdb.org/t/p/w154';
 const BACKDROP_PATH = 'http://image.tmdb.org/t/p/w1280';
 
-class MovieDetail extends PureComponent {
-	state = {
-		movie : {},
-	};
+const MovieDetail = () => {
+	// movies hook for state
+	const [ movie, setMovie ] = useState({
+		movie : [],
+	});
 
-	async componentDidMount() {
-		try {
-			const result = await fetch(
-				`https://api.themoviedb.org/3/movie/${this.props.match.params
-					.id}?api_key=3e11806009cadfb91187ad7b65b9dc21&language=en-US`,
-			);
-			const movie = await result.json();
+	// hook for api
+	useEffect(
+		() => {
+			axios
+				// get data from the API
+				.get(
+					`https://api.themoviedb.org/3/movie/${movie.id}?api_key=3e11806009cadfb91187ad7b65b9dc21&language=en-US`,
+				)
+				// set the results to state
+				.then((response) => {
+					console.log(response, 'response');
+					setMovie({ movie: response.data.results });
+				})
+				// catch errors
+				.catch((err) => console.log(err));
+		},
+		[ movie.id ],
+	);
 
-			this.setState({
-				movie : movie,
-			});
-		} catch (e) {
-			console.log(e);
-		}
-	}
+	return (
+		<MovieWrapper backdrop={`${BACKDROP_PATH}${movie.movie.backdrop_path}`}>
+			<MovieInfo>
+				<Link to='/'>
+					<Overdrive id={movie.id}>
+						<Poster
+							src={`${POSTER_PATH}${movie.poster_path}`}
+							alt={movie.title}
+						/>
+					</Overdrive>
+				</Link>
+				<div>
+					<h1>{movie.title}</h1>
+					<h3>{movie.release_date}</h3>
+					<p>{movie.overview}</p>
+				</div>
+			</MovieInfo>
+		</MovieWrapper>
+	);
+};
+// class MovieDetail extends PureComponent {
+// 	state = {
+// 		movie : {},
+// 	};
 
-	render() {
-		const { movie } = this.state;
-		return (
-			<MovieWrapper backdrop={`${BACKDROP_PATH}${movie.backdrop_path}`}>
-				<MovieInfo>
-					<Link to='/'>
-						<Overdrive id={movie.id}>
-							<Poster src={`${POSTER_PATH}${movie.poster_path}`} alt={movie.title} />
-						</Overdrive>
-					</Link>
-					<div>
-						<h1>{movie.title}</h1>
-						<h3>{movie.release_date}</h3>
-						<p>{movie.overview}</p>
-					</div>
-				</MovieInfo>
-			</MovieWrapper>
-		);
-	}
-}
+// 	async componentDidMount() {
+// 		try {
+// 			const result = await fetch(
+// 				`https://api.themoviedb.org/3/movie/${this.props.match.params
+// 					.id}?api_key=3e11806009cadfb91187ad7b65b9dc21&language=en-US`,
+// 			);
+// 			const movie = await result.json();
+
+// 			this.setState({
+// 				movie : movie,
+// 			});
+// 		} catch (e) {
+// 			console.log(e);
+// 		}
+// 	}
+
+// 	render() {
+// 		const { movie } = this.state;
+// 		return (
+// 			<MovieWrapper backdrop={`${BACKDROP_PATH}${movie.backdrop_path}`}>
+// 				<MovieInfo>
+// 					<Link to='/'>
+// 						<Overdrive id={movie.id}>
+// 							<Poster src={`${POSTER_PATH}${movie.poster_path}`} alt={movie.title} />
+// 						</Overdrive>
+// 					</Link>
+// 					<div>
+// 						<h1>{movie.title}</h1>
+// 						<h3>{movie.release_date}</h3>
+// 						<p>{movie.overview}</p>
+// 					</div>
+// 				</MovieInfo>
+// 			</MovieWrapper>
+// 		);
+// 	}
+// }
 
 export default MovieDetail;
